+++ test/e2e/specs/specs-calypso/wp-import-site-spec.js
