+++ test/e2e/specs/specs-calypso/wp-invite-users-spec.js
