+++ test/e2e/specs/specs-calypso/wp-launch-spec.js
