+++ test/e2e/specs/specs-calypso/wp-likes-spec.js
