+++ test/e2e/specs/specs-calypso/wp-log-in-out-spec.js
