+++ test/e2e/specs/specs-calypso/wp-manage-domains-spec.js
