+++ test/e2e/specs/specs-calypso/wp-media-editor-spec.js
