+++ test/e2e/specs/specs-calypso/wp-media-upload-spec.js
