+++ test/e2e/specs/specs-calypso/wp-reader-spec.js
