+++ test/e2e/specs/specs-calypso/wp-signup-spec.js
