+++ test/e2e/specs/specs-calypso/wp-stats-insights-spec.js
