+++ test/e2e/specs/specs-calypso/wp-theme-multisite-spec.js
