+++ test/e2e/specs/specs-calypso/wp-theme-switch-spec.js
