+++ test/e2e/specs/specs-wpcom/wp-calypso-gutenberg-focused-launch-spec.js
@@ -8,13 +8,13 @@ import { By, until } from 'selenium-webdriver';
 /**
  * Internal dependencies
  */
-import * as driverManager from '../lib/driver-manager.js';
-import * as dataHelper from '../lib/data-helper.js';
-import * as driverHelper from '../lib/driver-helper';
-import LoginFlow from '../lib/flows/login-flow.js';
-import DeleteSiteFlow from '../lib/flows/delete-site-flow.js';
-import GutenboardingFlow from '../lib/flows/gutenboarding-flow.js';
-import GutenbergEditorComponent from '../lib/gutenberg/gutenberg-editor-component';
+import * as driverManager from '../../lib/driver-manager.js';
+import * as dataHelper from '../../lib/data-helper.js';
+import * as driverHelper from '../../lib/driver-helper';
+import LoginFlow from '../../lib/flows/login-flow.js';
+import DeleteSiteFlow from '../../lib/flows/delete-site-flow.js';
+import GutenboardingFlow from '../../lib/flows/gutenboarding-flow.js';
+import GutenbergEditorComponent from '../../lib/gutenberg/gutenberg-editor-component';
 
 const mochaTimeOut = config.get( 'mochaTimeoutMS' );
 const startBrowserTimeoutMS = config.get( 'startBrowserTimeoutMS' );
