+++ test/e2e/specs/specs-wpcom/wp-calypso-gutenberg-page-editor-spec.js
@@ -7,21 +7,21 @@ import config from 'config';
 /**
  * Internal dependencies
  */
-import LoginFlow from '../lib/flows/login-flow.js';
+import LoginFlow from '../../lib/flows/login-flow.js';
 
-import ViewPagePage from '../lib/pages/view-page-page.js';
-import NotFoundPage from '../lib/pages/not-found-page.js';
+import ViewPagePage from '../../lib/pages/view-page-page.js';
+import NotFoundPage from '../../lib/pages/not-found-page.js';
 
-import PaypalCheckoutPage from '../lib/pages/external/paypal-checkout-page';
-import GutenbergEditorComponent from '../lib/gutenberg/gutenberg-editor-component';
-import GutenbergEditorSidebarComponent from '../lib/gutenberg/gutenberg-editor-sidebar-component';
-import PagePreviewComponent from '../lib/components/page-preview-component';
-import SimplePaymentsBlockComponent from '../lib/gutenberg/blocks/payment-block-component';
+import PaypalCheckoutPage from '../../lib/pages/external/paypal-checkout-page';
+import GutenbergEditorComponent from '../../lib/gutenberg/gutenberg-editor-component';
+import GutenbergEditorSidebarComponent from '../../lib/gutenberg/gutenberg-editor-sidebar-component';
+import PagePreviewComponent from '../../lib/components/page-preview-component';
+import SimplePaymentsBlockComponent from '../../lib/gutenberg/blocks/payment-block-component';
 
-import * as driverManager from '../lib/driver-manager.js';
-import * as mediaHelper from '../lib/media-helper.js';
-import * as dataHelper from '../lib/data-helper.js';
-import * as driverHelper from '../lib/driver-helper';
+import * as driverManager from '../../lib/driver-manager.js';
+import * as mediaHelper from '../../lib/media-helper.js';
+import * as dataHelper from '../../lib/data-helper.js';
+import * as driverHelper from '../../lib/driver-helper';
 
 const mochaTimeOut = config.get( 'mochaTimeoutMS' );
 const startBrowserTimeoutMS = config.get( 'startBrowserTimeoutMS' );
